changed: use the visualization manager in the visualization gui control
this unties the life cycle of a visualization from that of a
visualization control.

add new properties <slot> and <id> to the visualization control.
slot is a number identifying the 'slot' of the visualization to use.
this allows skins to run multiple visualizations simultanously.

id is the add-on id of the visualization to use. if it is blank,
the default is used.