changed: improve visualization handling.
Create() is called at the beginning of the life cycle, and Destroy()
at the end. These are called exactly once.
Start is called when a new rendering context for the visualization
is initialized, and Stop when the context is removed.

API is changed as the Start() now needs to be
passed screen info.