[cmake] mirror add-on builds in kodi build tree
this allows for easier development of add-ons on platforms
which do not depend on crosscompiling and which supplies
dependencies on the system level (typically linux).

by adding -DKODI_BUILD_DIR=<kodi build dir> to the add-on build,
it will mirror the binary and the addon-src/<addonid> folder in
addons/<addonid> in the kodi build tree.