[cmake] prepare a basic add-on build directory in the kodi build tree
this allows for easier development of add-ons on platforms
which do not depend on crosscompiling and which supplies
dependencies on the system level (typically linux).

by adding -DCMAKE_PREFIX_PATH=<kodi build dir>/build to the add-on build,
an add-on can be built easily against a specific version of kodi.